# backend/openapi.yaml
openapi: 3.0.3
info:
  title: EliteDial API
  description: Backend API for the EliteDial predictive dialer and call center platform.
  version: 1.0.0
servers:
  - url: /api
    description: Main API (all routes except webhooks)
  - url: /
    description: Root (health check and webhook endpoints)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, paused, completed, archived]
        dialMode:
          type: string
          enum: [predictive, progressive, preview]
        timezone:
          type: string
        maxAttemptsPerLead:
          type: integer
        abandonRateLimit:
          type: number
          format: float
        dialRatio:
          type: number
          format: float
        retryDelaySeconds:
          type: integer
        maxConcurrentCalls:
          type: integer
        aiTargetEnabled:
          type: boolean
        aiTarget:
          type: string
          nullable: true
        createdById:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CampaignList:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
        name:
          type: string
        sourceType:
          type: string
          enum: [upload, api, crm]
        uploadStatus:
          type: string
          enum: [pending, imported, failed]
        totalRecords:
          type: integer
        validRecords:
          type: integer
        duplicateSuppressed:
          type: integer
        dncSuppressed:
          type: integer
        invalidRecords:
          type: integer
        createdAt:
          type: string
          format: date-time
    CampaignContact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        primaryPhone:
          type: string
        email:
          type: string
          nullable: true
        status:
          type: string
          enum: [queued, dialing, connected, completed, suppressed-dnc, failed]
        priority:
          type: integer
        attemptCount:
          type: integer
    ImportResult:
      type: object
      properties:
        listId:
          type: string
          format: uuid
        totalRecords:
          type: integer
        validRecords:
          type: integer
        invalidRecords:
          type: integer
        duplicateSuppressed:
          type: integer
        dncSuppressed:
          type: integer
    Paginated:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
security:
  - bearerAuth: []
paths:
...
  /campaigns:
    get:
      tags: [Campaigns]
      summary: List all campaigns (supervisor+)
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
    post:
      tags: [Campaigns]
      summary: Create a new campaign (supervisor+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                dialMode:
                  type: string
                  enum: [predictive, progressive, preview]
                timezone:
                  type: string
                maxAttemptsPerLead:
                  type: integer
                abandonRateLimit:
                  type: number
                dialRatio:
                  type: number
                retryDelaySeconds:
                  type: integer
                maxConcurrentCalls:
                  type: integer
                aiTargetEnabled:
                  type: boolean
                aiTarget:
                  type: string
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
  /campaigns/{id}:
    get:
      tags: [Campaigns]
      summary: Get campaign details (supervisor+)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign with lists and counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Campaigns]
      summary: Update a campaign (supervisor+)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                dialMode:
                  type: string
                  enum: [predictive, progressive, preview]
                timezone:
                  type: string
                maxAttemptsPerLead:
                  type: integer
                abandonRateLimit:
                  type: number
                dialRatio:
                  type: number
                retryDelaySeconds:
                  type: integer
                maxConcurrentCalls:
                  type: integer
                aiTargetEnabled:
                  type: boolean
                aiTarget:
                  type: string
      responses:
        '200':
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
  /campaigns/{id}/start:
    post:
      tags: [Campaigns]
      summary: Start a campaign (supervisor+)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
  /campaigns/{id}/pause:
    post:
      tags: [Campaigns]
      summary: Pause a campaign (supervisor+)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
  /campaigns/{id}/contacts:
    get:
      tags: [Campaigns]
      summary: List contacts in a campaign (supervisor+, paginated)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated contacts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties:
                      contacts:
                        type: array
                        items:
                          $ref: '#/components/schemas/CampaignContact'
  /campaigns/{id}/lists:
    post:
      tags: [Campaigns]
      summary: Create a contact list in a campaign (supervisor+)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                sourceType:
                  type: string
                  enum: [upload, api, crm]
      responses:
        '201':
          description: List created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignList'
  /campaigns/{id}/lists/{listId}/import:
    post:
      tags: [Campaigns]
      summary: Import contacts into a campaign list (supervisor+)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: listId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rows:
                  type: array
                  items:
                    type: object
                    properties:
                      firstName:
                        type: string
                      lastName:
                        type: string
                      phone:
                        type: string
                      email:
                        type: string
                      timezone:
                        type: string
                      priority:
                        type: integer
                csv:
                  type: string
                  description: Raw CSV text (alternative to rows)
      responses:
        '200':
          description: Import results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
  /campaigns/{id}/import:
    post:
      tags: [Campaigns]
      summary: Quick-import contacts with auto-created list (supervisor+)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                listName:
                  type: string
                rows:
                  type: array
                  items:
                    type: object
                csv:
                  type: string
      responses:
        '201':
          description: Import results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
...